// frontend/src/company.d.ts
export interface Doctor {
  DoctorID: number;
  Name: string;
  Specialization: string;
  PhoneNumber: string;
  DepartmentID: number;
}